+++ src/main/java/com/clinicappbdas2/model/views/NavstevaPacienta.java
@@ -6,17 +6,17 @@ import java.util.Date;
 
 @Data
 public class NavstevaPacienta {
-    private Integer idNavsteva;
-    private Date datum;
-    private Integer idPacient;
-    private Integer idZamestnanec;
-    private String problem;
-    private String rekomendace;
-    private Integer idStatus;
-    private String pacientJmeno;
-    private String pacientPrijmeni;
-    private Integer cisloTelefonu;
-    private String zamestnanecJmeno;
-    private String zamestnanecPrijmeni;
-    private String status;
+    private Long        idNavsteva;
+    private Date        datum;
+    private Long        idPacient;
+    private Long        idZamestnanec;
+    private String      problem;
+    private String      rekomendace;
+    private Long        idStatus;
+    private String      pacientJmeno;
+    private String      pacientPrijmeni;
+    private Long        cisloTelefonu;
+    private String      zamestnanecJmeno;
+    private String      zamestnanecPrijmeni;
+    private String      status;
 }
