+++ src/main/java/com/clinicappbdas2/model/views/ZamestnanecData.java
@@ -6,17 +6,17 @@ import java.util.Date;
 
 @Data
 public class ZamestnanecData {
-    private Integer idZamestnanec;
+    private Long idZamestnanec;
     private String jmeno;
     private String prijmeni;
     private Date datumNarozeni;
-    private Integer cisloTelefonu;
-    private Integer pracovniZkusenosti;
-    private Integer idAdresa;
+    private Long cisloTelefonu;
+    private Long pracovniZkusenosti;
+    private Long idAdresa;
     private String zeme;
     private String mesto;
     private String adresa;
-    private Integer psc;
+    private Long psc;
     private Integer idOddeleni;
     private String nazevOddeleni;
